Added field for server cancellation date. Rearranged dates on server info page in the process.
The cancellation date can be set in advance to know when the server will no
longer have to be paid for and when it will be terminated by the provider.
Cancelled servers will appear greyed out in the server list (once the
cancellation date has passed) and will have their cancellation date listed
instead of their next due date (when the next due date would be after the
cancellation date).